+++ NEWS.md
@@ -1,4 +1,12 @@
-# shinyLP 0.1.0
+# shinyLP 1.1.0
+
+* fixing mistakes from last release for description author field and adjusted thumbnail label component sizing
+* adding new functions for: iframe, list group, wells
+* additional examples on the sample shiny app
+* now making use of `shinythemes` package instead of direct link to css page
+
+
+# shinyLP 1.0.0
 
 * now on CRAN https://CRAN.R-project.org/package=shinyLP - 09/13/2016
 * making changes suggested from win-build